requirements.txt
python-dotenv
ShopifyAPI
ipykernel